+++ src/geom_core/XSecSurf.h
@@ -63,7 +63,7 @@ public:
     int GetPrincipalDirection() { return m_PrincipalDir; }
 
     void GetBasicTransformation( double w, Matrix4d &mat );
-    static void GetBasicTransformation( int pdir, int wdir, int wshift, bool flip, double w, Matrix4d &mat );
+    static void GetBasicTransformation( int pdir, int wdir, double wshift, bool flip, double w, Matrix4d &mat );
 
     void SetGlobalXForm( const Matrix4d &mat )            { m_GlobalXForm = mat; }
     Matrix4d GetGlobalXForm()                            { return m_GlobalXForm; }
